remove old timeline window
